Part of the web agregator implemented.
Starting from the list of member's websites, it load the html content
of the site, search for a feed url, load the feed using Rome e present
the feed in the user interface. In the user interface, it's possible to
reference the website in order to start collecting content from its
feed.